CORE-345: (Re)Implement Transfer, Wallet, WalletManager using BRCrypto* Core
