+++ .gitignore
@@ -39,7 +39,3 @@
 
 test
 .git-re*
-Swift/BRCrypto/BRCryptoA.swift
-Swift/BRCrypto/BRCryptoB.swift
-crypto/BRCryptoAmount.c
-crypto/BRCryptoNetwork.c
