+++ Java/.idea/modules.xml
@@ -4,9 +4,12 @@
     <modules>
       <module fileurl="file://$PROJECT_DIR$/Core/Core.iml" filepath="$PROJECT_DIR$/Core/Core.iml" />
       <module fileurl="file://$PROJECT_DIR$/CoreCrypto/CoreCrypto.iml" filepath="$PROJECT_DIR$/CoreCrypto/CoreCrypto.iml" />
+      <module fileurl="file://$PROJECT_DIR$/CoreCrypto/CoreCrypto.iml" filepath="$PROJECT_DIR$/CoreCrypto/CoreCrypto.iml" />
       <module fileurl="file://$PROJECT_DIR$/CoreDemo/CoreDemo.iml" filepath="$PROJECT_DIR$/CoreDemo/CoreDemo.iml" />
       <module fileurl="file://$PROJECT_DIR$/CoreNative/CoreNative.iml" filepath="$PROJECT_DIR$/CoreNative/CoreNative.iml" />
+      <module fileurl="file://$PROJECT_DIR$/CoreNative/CoreNative.iml" filepath="$PROJECT_DIR$/CoreNative/CoreNative.iml" />
       <module fileurl="file://$PROJECT_DIR$/Crypto/Crypto.iml" filepath="$PROJECT_DIR$/Crypto/Crypto.iml" />
+      <module fileurl="file://$PROJECT_DIR$/CryptoDemo/CryptoDemo.iml" filepath="$PROJECT_DIR$/CryptoDemo/CryptoDemo.iml" />
       <module fileurl="file://$PROJECT_DIR$/CryptoDemo/CryptoDemo.iml" filepath="$PROJECT_DIR$/CryptoDemo/CryptoDemo.iml" />
       <module fileurl="file://$PROJECT_DIR$/Java.iml" filepath="$PROJECT_DIR$/Java.iml" />
     </modules>
